+++ contracts/contract/src/storage_keys.rs
@@ -1,3 +1,4 @@
 pub const PROTOCOL_DEBTS_KEY: &str = "pd";
 pub const MARGIN_BASE_TOKEN_LIMIT: &str = "bl";
 pub const MARGIN_BASE_TOKEN_LIMIT_GUR: &str = "bg";
+pub const CLIENT_ECHO_SENDER_WHITELIST: &str = "csw";